Correcting errors in the register package and testing the program's functionality in Swagger

+++ src/main/java/com/Kronos/Kronos/Register/service/UserService.java
@@ -50,7 +50,10 @@ public class UserService {
     }
 
     public void deleteUser(Long id) {
-        userRepository.deleteById(id);
+        if (!userRepository.existsById(id)) {
+            throw new RuntimeException("Usuário não encontrado");
+        } else {
+            userRepository.deleteById(id);
     }
 
 }
